refactor(49): refactor package name
